feat: defered configuration
This allows neovim to process the UI before configuring some chunk of plugins, making it so
the file is visible while the plugins are loading. The user still cannot interact with neovim
in during the loading, but having the contents displayed allows me to start scanning the file
to decide where to jump to while it happens. It gives only the impression of a fast startup,
but I find it better than looking at a blank screen for half a second in some cases. At least
I have information to start thinking about my next step while it loads.