fix: change path_display to show full path
Smart is actually pretty bad, it shortens the path to the difference, so
in projects with many levels of depth you cannot know anymore where the
file actually is. "filename_first" show the full relative path after the
filename, so you can actually know where files with same name are deep
in the tree.