fix: remove vim.schedule to fix whichkey on fish files
I still don't know why WhichKey would fail to load properly (without
errors) on fish files (and I'm sure I've seen it fail before, but I
don't remember what file it was) if I use the vim.schedule trick. But it
seems that the trick is more placebo than anything, so I'm removing it.
The only difference I'm noticing today is that the UI loads at once at
the end instead of flashing black and then getting themed, with the
overall time-to-ready/fully-visible-and-themed/be-able-to-interact.