refactor: many small changes, specially in UI
The initial idea was to review the plugins and reorganize the
configuration. However, I find it hard to remove plugins as there is
nothing really useless and the configuration was already quite tidy,
with only a few changes to remove a small clutter, which came mostly
from me not wanting to create files for simple configurations. The UI
was changed for catppuccin and to better integrate with my current tmux
setup, so there are some visual changes.