fix(fold): add a fold toggle to fix slowness in big markdown files
The foldexpr set to the tree-sitter function causes a lot of slowdown
when used in big markdown files. This toggle sets it to manual, so
tree-sitter stops processing it.

However, just setting the option is not enough for some reason, so I
open and close the help window because that forces the options to apply.